move and refactor comment callback functions
