+++ packages/lesswrong/server/languageModels/modGPT.ts
@@ -1,27 +1,15 @@
 import { getOpenAI } from './languageModelIntegration';
-import type { UpdateCallbackProperties } from '../mutationCallbacks';
 import { isAnyTest } from '../../lib/executionEnvironment';
 import sanitizeHtml from 'sanitize-html';
 import { sanitizeAllowedTags } from '../../lib/vulcan-lib/utils';
-import { htmlToText } from 'html-to-text';
-import { createMutator, updateMutator } from '../vulcan-lib/mutators';
+import { updateMutator } from '../vulcan-lib/mutators';
 import Comments from '../../lib/collections/comments/collection';
-import Posts from '../../lib/collections/posts/collection';
-import { EA_FORUM_COMMUNITY_TOPIC_ID } from '../../lib/collections/tags/helpers';
 import { dataToHTML } from '../editor/conversionUtils';
-import { isEAForum } from '../../lib/instanceSettings';
-import Users from '../../lib/collections/users/collection';
-import { commentGetPageUrlFromIds } from '../../lib/collections/comments/helpers';
 import OpenAI from 'openai';
-import Conversations from '../../lib/collections/conversations/collection';
-import Messages from '../../lib/collections/messages/collection';
-import { getAdminTeamAccount } from '../callbacks/commentCallbacks';
 import { captureEvent } from '../../lib/analyticsEvents';
-import { appendToSunshineNotes } from '../../lib/collections/users/helpers';
-import { createAdminContext } from "../vulcan-lib/query";
 import difference from 'lodash/difference';
 import { truncatise } from '../../lib/truncatise';
-import { FetchedFragment, fetchFragmentSingle } from '../fetchFragment';
+import { FetchedFragment } from '../fetchFragment';
 
 
 export const modGPTPrompt = `
@@ -106,7 +94,7 @@ export const sanitizeHtmlOptions = {
 /**
  * Ask GPT-4 to help moderate the given comment. It will respond with a "recommendation", as per the prompt above.
  */
-async function checkModGPT(comment: DbComment, post: FetchedFragment<'PostsOriginalContents'>): Promise<void> {
+export async function checkModGPT(comment: DbComment, post: FetchedFragment<'PostsOriginalContents'>): Promise<void> {
   const api = await getOpenAI();
   if (!api) {
     if (!isAnyTest) {
@@ -260,80 +248,4 @@ async function checkModGPT(comment: DbComment, post: FetchedFragment<'PostsOrigi
     }
     return
   }
-}
-
-// TODO: move this to a commentCallbackFunctions file.  This was an updateAsync.
-async function checkModGPTOnCommentUpdate({oldDocument, newDocument}: UpdateCallbackProperties<"Comments">) {
-  // On the EA Forum, ModGPT checks earnest comments on posts for norm violations.
-  // We skip comments by unreviewed authors, because those will be reviewed by a human.
-  if (
-    !isEAForum ||
-    !newDocument.postId ||
-    newDocument.deleted ||
-    newDocument.deletedPublic ||
-    newDocument.spam ||
-    newDocument.needsReview ||
-    newDocument.authorIsUnreviewed ||
-    newDocument.retracted ||
-    newDocument.rejected ||
-    newDocument.shortform ||
-    newDocument.moderatorHat ||
-    !newDocument.contents?.originalContents?.data
-  ) {
-    return
-  }
-  
-  const noChange = oldDocument.contents?.originalContents?.data === newDocument.contents.originalContents.data
-  if (noChange) return
-
-  // only have ModGPT check comments on posts tagged with "Community"
-  const post = await fetchFragmentSingle({
-    collectionName: "Posts",
-    fragmentName: "PostsOriginalContents",
-    currentUser: null,
-    skipFiltering: true,
-    selector: {_id: newDocument.postId},
-  });
-  if (!post) return
-  
-  const postTags = post.tagRelevance
-  if (!postTags || !Object.keys(postTags).includes(EA_FORUM_COMMUNITY_TOPIC_ID)) return
-  
-  void checkModGPT(newDocument, post)
-}
-
-// TODO: move this to a commentCallbackFunctions file.  This was a createAsync.
-async function checkModGPTOnCommentCreate({document}: {document: DbComment}) {
-  // On the EA Forum, ModGPT checks earnest comments on posts for norm violations.
-  // We skip comments by unreviewed authors, because those will be reviewed by a human.
-  if (
-    !isEAForum ||
-    !document.postId ||
-    document.deleted ||
-    document.deletedPublic ||
-    document.spam ||
-    document.needsReview ||
-    document.authorIsUnreviewed ||
-    document.retracted ||
-    document.rejected ||
-    document.shortform ||
-    document.moderatorHat
-  ) {
-    return
-  }
-  
-  // only have ModGPT check comments on posts tagged with "Community"
-  const post = await fetchFragmentSingle({
-    collectionName: "Posts",
-    fragmentName: "PostsOriginalContents",
-    currentUser: null,
-    skipFiltering: true,
-    selector: {_id: document.postId},
-  });
-  if (!post) return
-  
-  const postTags = post.tagRelevance
-  if (!postTags || !Object.keys(postTags).includes(EA_FORUM_COMMUNITY_TOPIC_ID)) return
-  
-  void checkModGPT(document, post)
 }
