+++ packages/lesswrong/server/mutationCallbacks.ts
@@ -50,13 +50,13 @@ export class CollectionMutationCallbacks<N extends CollectionNameString> {
   newAfter: CallbackChainHook<ObjectsByCollectionName[N],[DbUser|null, AfterCreateCallbackProperties<N>]>
 
   createAsync: CallbackHook<[AfterCreateCallbackProperties<N>]>
-  newAsync: CallbackHook<[ObjectsByCollectionName[N],DbUser|null,any]>
+  newAsync: CallbackHook<[ObjectsByCollectionName[N],DbUser|null,any,AfterCreateCallbackProperties<N>]>
 
   updateValidate: CallbackChainHook<CallbackValidationErrors,[UpdateCallbackProperties<N>]>
 
   updateBefore: CallbackChainHook<Partial<ObjectsByCollectionName[N]>,[UpdateCallbackProperties<N>]>
 
-  editSync: CallbackChainHook<MongoModifier<ObjectsByCollectionName[N]>,[ObjectsByCollectionName[N],DbUser|null,ObjectsByCollectionName[N]]>
+  editSync: CallbackChainHook<MongoModifier<ObjectsByCollectionName[N]>,[ObjectsByCollectionName[N],DbUser|null,ObjectsByCollectionName[N],UpdateCallbackProperties<N>]>
 
   updateAfter: CallbackChainHook<ObjectsByCollectionName[N],[UpdateCallbackProperties<N>]>
 
@@ -75,11 +75,11 @@ export class CollectionMutationCallbacks<N extends CollectionNameString> {
     this.createAfter = new CallbackChainHook<ObjectsByCollectionName[N],[AfterCreateCallbackProperties<N>]>(`${namePrefix}.create.after`);
     this.newAfter = new CallbackChainHook<ObjectsByCollectionName[N],[DbUser|null, AfterCreateCallbackProperties<N>]>(`${namePrefix}.new.after`);
     this.createAsync = new CallbackHook<[AfterCreateCallbackProperties<N>]>(`${namePrefix}.create.async`);
-    this.newAsync = new CallbackHook<[ObjectsByCollectionName[N],DbUser|null,any]>(`${namePrefix}.new.async`);
+    this.newAsync = new CallbackHook<[ObjectsByCollectionName[N],DbUser|null,any,AfterCreateCallbackProperties<N>]>(`${namePrefix}.new.async`);
 
     this.updateValidate = new CallbackChainHook<CallbackValidationErrors,[UpdateCallbackProperties<N>]>(`${namePrefix}.update.validate`);
     this.updateBefore = new CallbackChainHook<Partial<ObjectsByCollectionName[N]>,[UpdateCallbackProperties<N>]>(`${namePrefix}.update.before`);
-    this.editSync = new CallbackChainHook<MongoModifier<ObjectsByCollectionName[N]>,[ObjectsByCollectionName[N],DbUser|null,ObjectsByCollectionName[N]]>(`${namePrefix}.edit.sync`);
+    this.editSync = new CallbackChainHook<MongoModifier<ObjectsByCollectionName[N]>,[ObjectsByCollectionName[N],DbUser|null,ObjectsByCollectionName[N],UpdateCallbackProperties<N>]>(`${namePrefix}.edit.sync`);
     this.updateAfter = new CallbackChainHook<ObjectsByCollectionName[N],[UpdateCallbackProperties<N>]>(`${namePrefix}.update.after`);
     this.updateAsync = new CallbackHook<[UpdateCallbackProperties<N>]>(`${namePrefix}.update.async`);
     this.editAsync = new CallbackHook<[ObjectsByCollectionName[N],ObjectsByCollectionName[N],DbUser|null,CollectionBase<N>,UpdateCallbackProperties<N>]>(`${namePrefix}.edit.async`)
