+++ packages/lesswrong/server/resolvers/alignmentForumMutations.ts
@@ -4,7 +4,7 @@ import { userCanMakeAlignmentPost } from '../../lib/alignment-forum/users/helper
 import { accessFilterSingle } from '../../lib/utils/schemaUtils';
 import { moveToAFUpdatesUserAFKarma } from '../callbacks/alignment-forum/callbacks';
 import { postsMoveToAFAddsAlignmentVoting } from '../callbacks/alignment-forum/alignmentPostCallbacks';
-import { commentsAlignmentEdit, recalculateAFCommentMetadata } from '../callbacks/alignment-forum/alignmentCommentCallbacks';
+import { commentsAlignmentEdit, recalculateAFCommentMetadata } from '../callbacks/commentCallbackFunctions';
 
 const alignmentCommentResolvers = {
   Mutation: {
@@ -17,8 +17,8 @@ const alignmentCommentResolvers = {
         await context.Comments.rawUpdateOne({_id: commentId}, modifier);
         const updatedComment = (await context.Comments.findOne(commentId))!
         await moveToAFUpdatesUserAFKarma(updatedComment, comment);
-        await recalculateAFCommentMetadata(comment.postId)
-        await commentsAlignmentEdit(updatedComment, comment);
+        await recalculateAFCommentMetadata(comment.postId, context);
+        await commentsAlignmentEdit(updatedComment, comment, context);
         return await accessFilterSingle(context.currentUser, context.Comments, updatedComment, context);
       } else {
         throw new Error({id: `app.user_cannot_edit_comment_alignment_forum_status`} as any);
