+++ packages/lesswrong/server/resolvers/conversationResolvers.ts
@@ -1,5 +1,5 @@
 import { forumSelect } from "@/lib/forumTypeUtils";
-import { getAdminTeamAccount } from "../callbacks/commentCallbacks";
+import { getAdminTeamAccount } from "../utils/adminTeamAccount";
 import { defineMutation } from "../utils/serverGraphqlUtil";
 import { TupleSet, UnionOf } from "@/lib/utils/typeGuardUtils";
 import { createMutator } from "../vulcan-lib/mutators";
