+++ packages/lesswrong/server/utils/callbackHooks.ts
@@ -6,6 +6,8 @@ type MaybePromise<T> = T|Promise<T>
 
 export type CallbackChainFn<IteratorType, ArgumentsType extends any[]> = (doc: IteratorType, ...args: ArgumentsType) => (MaybePromise<IteratorType> | undefined | void)
 
+const LOG_ADDED_CALLBACK_HOOK_NAME = 'comments.';
+
 /**
  * A set of callbacks which run in a chain, each modifying a value and passing
  * it to the next. Each callback in the chain receives a current-version of the
@@ -29,6 +31,14 @@ export class CallbackChainHook<IteratorType,ArgumentsType extends any[]> {
   
   add = (fn: CallbackChainFn<IteratorType,ArgumentsType>) => {
     this._callbacks.push(fn);
+    if (this._name.startsWith(LOG_ADDED_CALLBACK_HOOK_NAME)) {
+      if (fn.name) {
+        console.log(`${this._name} added callback ${fn.name}`);
+      } else {
+        console.log(`${this._name} added callback without name`);
+        console.log(new Error().stack?.split('\n').slice(2, 3)[0]);
+      }
+    }
     if (this._callbacks.length > 20) {
       // eslint-disable-next-line no-console
       console.log(`Warning: Excessively many callbacks (${this._callbacks.length}) on hook ${this._name}.`);
@@ -119,6 +129,14 @@ export class CallbackHook<ArgumentsType extends any[]> {
   
   add = (fn: (...args: ArgumentsType) => void|Promise<void>) => {
     this._callbacks.push(fn);
+    if (this._name.startsWith(LOG_ADDED_CALLBACK_HOOK_NAME)) {
+      if (fn.name) {
+        console.log(`${this._name} added callback ${fn.name}`);
+      } else {
+        console.log(`${this._name} added callback without name`);
+        console.log(new Error().stack?.split('\n').slice(2, 3)[0]);
+      }
+    }
     if (this._callbacks.length > 20) {
       // eslint-disable-next-line no-console
       console.log(`Warning: Excessively many callbacks (${this._callbacks.length}) on hook ${this._name}.`);
